+++ browser-tests/utils/getDropdownOption.ts
@@ -1,0 +1,7 @@
+import { screen } from '@testing-library/testcafe';
+
+function getDropdownOption(label: string) {
+  return screen.getByRole('option', { name: label });
+}
+
+export default getDropdownOption;