+++ src/common/test/TestProviders.tsx
@@ -3,6 +3,7 @@ import { Provider } from 'react-redux';
 import { MockedProvider, MockedResponse } from '@apollo/client/testing';
 import { createBrowserHistory } from 'history';
 import { Router } from 'react-router';
+import { HelmetProvider } from 'react-helmet-async';
 
 import { store } from '../../domain/app/state/AppStore';
 
@@ -15,7 +16,9 @@ const TestProviders = ({ children, mocks }: Props) => {
   return (
     <Provider store={store}>
       <MockedProvider mocks={mocks} addTypename={false}>
-        <Router history={createBrowserHistory()}>{children}</Router>
+        <HelmetProvider>
+          <Router history={createBrowserHistory()}>{children}</Router>
+        </HelmetProvider>
       </MockedProvider>
     </Provider>
   );